// src/features/creative-space/types.ts
export enum NodeType {
  PROMPT_INPUT = 'PROMPT_INPUT',
  IMAGE_GENERATOR = 'IMAGE_GENERATOR',
  VIDEO_GENERATOR = 'VIDEO_GENERATOR',
  VIDEO_ANALYZER = 'VIDEO_ANALYZER',
  IMAGE_EDITOR = 'IMAGE_EDITOR',
  AUDIO_GENERATOR = 'AUDIO_GENERATOR',
  SCRIPT_PLANNER = 'SCRIPT_PLANNER',
  SCRIPT_EPISODE = 'SCRIPT_EPISODE',
  STORYBOARD_GENERATOR = 'STORYBOARD_GENERATOR',
  STORYBOARD_IMAGE = 'STORYBOARD_IMAGE',
  STORYBOARD_SPLITTER = 'STORYBOARD_SPLITTER',
  CHARACTER_NODE = 'CHARACTER_NODE',
  DRAMA_ANALYZER = 'DRAMA_ANALYZER',
  DRAMA_REFINED = 'DRAMA_REFINED',
  STYLE_PRESET = 'STYLE_PRESET',
  SORA_VIDEO_GENERATOR = 'SORA_VIDEO_GENERATOR',
  SORA_VIDEO_CHILD = 'SORA_VIDEO_CHILD',
  STORYBOARD_VIDEO_GENERATOR = 'STORYBOARD_VIDEO_GENERATOR',
  STORYBOARD_VIDEO_CHILD = 'STORYBOARD_VIDEO_CHILD',
  VIDEO_EDITOR = 'VIDEO_EDITOR',
}